+++ app/(auth)/forgot-password.tsx
@@ -1,16 +1,17 @@
+import AuthPageLayout, { authLayoutStyles } from "@/components/auth/AuthPageLayout";
 import Button from "@/components/ui/button";
 import TextInput from "@/components/ui/text-input";
 import Colors from "@/constants/Colors";
 import { HttpWrapper } from "@/shared-libs/utils/http-wrapper";
 import Toaster from "@/shared-uis/components/toaster/Toaster";
-import fnStyles from "@/styles/forgot-password.styles";
+import fnStyles from "@/styles/signup.styles";
 import { useTheme } from "@react-navigation/native";
 import { useRouter } from "expo-router";
-import * as React from "react";
+import React, { useState } from "react";
 import { Image, Text, View } from "react-native";
 
 const ForgotPasswordScreen = () => {
-    const [email, setEmail] = React.useState("");
+    const [email, setEmail] = useState("");
     const router = useRouter();
     const theme = useTheme();
     const styles = fnStyles(theme);
@@ -57,49 +58,52 @@ const ForgotPasswordScreen = () => {
     };
 
     return (
-        <View style={styles.container}>
-            {/* Logo Section */}
+        <AuthPageLayout>
             <Image
-                source={require("@/assets/images/logo.png")} // Replace with your actual logo path
+                source={require("@/assets/images/logo.png")}
                 style={styles.logo}
                 resizeMode="contain"
             />
-
-            {/* Title */}
-            <Text style={styles.title}>Forgot Password</Text>
-
-            {/* Email Input Field */}
-            <TextInput
-                autoCapitalize="none"
-                label="Enter your Email ID"
-                value={email}
-                onChangeText={setEmail}
-                mode="outlined"
-                style={styles.input}
-                theme={{ colors: { primary: Colors(theme).text } }}
-            />
-
-            {/* Reset Password Button */}
-            <Button
-                mode="contained"
-                onPress={() => {
-                    handleResetPassword();
-                }}
-            >
-                Reset Password
-            </Button>
-
-            {/* Back to Login Prompt */}
-            <Text style={styles.loginText}>
-                Remember your password?{" "}
-                <Text
-                    style={styles.loginLink}
+            <View style={authLayoutStyles.formHeader}>
+                <Text style={[styles.title, authLayoutStyles.formTitle]}>
+                    Forgot Password
+                </Text>
+                <Text style={[styles.subTitle, authLayoutStyles.formSubtitle]}>
+                    We will email you a reset link.
+                </Text>
+            </View>
+            <View style={[styles.inputContainer, authLayoutStyles.inputStack]}>
+                <TextInput
+                    autoCapitalize="none"
+                    label="Enter your Email ID"
+                    value={email}
+                    onChangeText={setEmail}
+                    mode="outlined"
+                    textColor={Colors(theme).text}
+                    style={styles.input}
+                    theme={{ colors: { primary: Colors(theme).text } }}
+                />
+                <Button
+                    mode="contained"
+                    style={authLayoutStyles.primaryButton}
+                    onPress={handleResetPassword}
+                >
+                    Reset Password
+                </Button>
+            </View>
+            <View style={authLayoutStyles.loginPrompt}>
+                <Text style={[styles.loginText, authLayoutStyles.loginText]}>
+                    Remember your password?
+                </Text>
+                <Button
+                    mode="outlined"
+                    style={authLayoutStyles.secondaryButton}
                     onPress={() => router.replace("/(auth)/login")}
                 >
                     Login
-                </Text>
-            </Text>
-        </View>
+                </Button>
+            </View>
+        </AuthPageLayout>
     );
 };
